finished route refactoring
